+++ frontend/src/pages/ExploreData/ExploreDataPage.module.scss
